Image coreg changes - modified job create struct to only contain trigger Url as it seems the job will have to download the initial json too. Coreg processing is now in its own package.
